delete api and upload img api

+++ Service/Service.cs
@@ -1,25 +1,32 @@
 ﻿using System.Net;
 using E_commerce.Server.DAL.BASE;
+using E_commerce.Server.data;
 using E_commerce.Server.Model.DTO;
 using E_commerce.Server.Model.Entities;
 using Microsoft.AspNetCore.Mvc;
+using static System.Reflection.Metadata.BlobBuilder;
 
 namespace E_commerce.Server.Service
 {
     public class Service : IService
     {
         private readonly IRepository<Books> _booksRepository;
+        private readonly IRepository<BookImg> _bookImgRepository;
+        private readonly ApplicationDbContext _DbContext;
 
-        public Service(IRepository<Books> booksRepository)
+        public Service(IRepository<Books> booksRepository ,IRepository<BookImg> BookImgRepo  ,ApplicationDbContext dbContext)
         {
             _booksRepository = booksRepository;
+            _DbContext = dbContext;
+            _bookImgRepository = BookImgRepo;
         }
 
-        public async Task<(int statusCode, IEnumerable<Books>? Books, bool success)> GetBooks()
+        public async Task<(int statusCode, IEnumerable<BookRes>? Books, bool success)> GetBooks()
         {
             try
             {
                 var books = await _booksRepository.GetAll();
+                var bookImg = await _bookImgRepository.GetAll();
 
                 if (books == null || !books.Any())
                 {
@@ -36,7 +43,24 @@ namespace E_commerce.Server.Service
                     return (404, null, false);
                 }
 
-                return (200, books, true);
+
+                var data = (from bookdetail in books
+                            join img in bookImg
+                            on bookdetail.Book_Id equals img.Book_id into bookImages
+                            from img in bookImages.DefaultIfEmpty()
+                            select new BookRes
+                            {
+                                Book_Id = bookdetail.Book_Id,
+                                Title = bookdetail.Title,
+                                Author = bookdetail.Author,
+                                ISBN = bookdetail.ISBN,
+                                Quantity = bookdetail.Quantity,
+                                IsDeleted = bookdetail.IsDeleted,
+                                FilePath = img != null ? img.FilePath : null
+                            }).ToList();
+
+
+                return (200, data , true);
             }
             catch
             {
@@ -45,11 +69,13 @@ namespace E_commerce.Server.Service
         }
 
 
-        public async Task<(int statusCode, IEnumerable<Books>? Books, bool success)> getDeleteBooks()
+        public async Task<(int statusCode, IEnumerable<BookRes>? Books, bool success)> getDeleteBooks()
         {
             try
             {
                 var books = await _booksRepository.GetAll();
+
+                var bookImg = await _bookImgRepository.GetAll();
 
                 if (books == null || !books.Any())
                 {
@@ -70,7 +96,24 @@ namespace E_commerce.Server.Service
                     return (404, null, false);
                 }
 
-                return (200, books, true);
+
+                var data = (from bookdetail in books
+                            join img in bookImg
+                            on bookdetail.Book_Id equals img.Book_id into bookImages
+                            from img in bookImages.DefaultIfEmpty()
+                            select new BookRes
+                            {
+                                Book_Id = bookdetail.Book_Id,
+                                Title = bookdetail.Title,
+                                Author = bookdetail.Author,
+                                ISBN = bookdetail.ISBN,
+                                Quantity = bookdetail.Quantity,
+                                IsDeleted = bookdetail.IsDeleted,
+                                FilePath = img != null ? img.FilePath : null
+                            }).ToList();
+        
+
+                return (200, data, true);
             }
             catch
             {
@@ -155,12 +198,23 @@ namespace E_commerce.Server.Service
             try
             {
                 var book = await _booksRepository.GetById(book_id);
+
+                var BookImg = await _bookImgRepository.GetAll();
+
+                var img = BookImg.FirstOrDefault(b => b.Book_id == book_id);
+
+                if (img == null)
+                {
+                    return (404, false);
+                }
+
                 if (book == null)
                 {
                     return (404, false);
                 }
 
                 await _booksRepository.Delete(book);
+                await _bookImgRepository.Delete(img);
                 return (200, true);
             }
             catch
@@ -168,6 +222,14 @@ namespace E_commerce.Server.Service
                 return (500, false);
             }
         }
+
+
+
+
+
+
+
+
 
         public async Task<(int StatusCode, IEnumerable<Books>? Books, bool success)> GetById(int book_id)
         {
